packages/api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
sqlalchemy==2.0.25
asyncpg==0.29.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.13.1
email-validator==2.1.0.post1
aiofiles==23.2.1
httpx==0.26.0
openai==1.12.0
pgvector==0.2.4
psycopg2-binary==2.9.9
python-dotenv==1.0.1